Refactor task monitor view layout to FXML
